// src/enum.ts
export enum EBlockKey {
    X = "X",
    Category = "category",
    Y = "Y",
}


export enum EViewOption {
    GridX = "grid_x",
    GridY = "grid_y",
    PointerStyle = "pointer_style",
    Legend = "liner",
    Line = "line",
    Fill = "fill",
    LineTension = "line_tension",
    DataLabels = "DataLabels",
    TickX = "TickX",
    TickY = "TickY",
}